requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
timeago==1.0.16